# backend/src/swagger.yaml
openapi: 3.0.3
info:
  title: JMeter Training App API
  description: >
    A simple task management API designed as a target for JMeter performance
    testing training. Includes intentionally slow and flaky endpoints to
    practice load testing, error handling, and assertions.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Native (local)
  - url: http://localhost
    description: Docker (via nginx)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: Internal server error
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account. Username must be unique and max 100 characters.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  maxLength: 100
                  example: perf_user_1
                password:
                  type: string
                  example: TestPass123!
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Account created successfully
        "400":
          description: Missing fields or username too long
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/login:
    post:
      summary: Login
      description: >
        Authenticates a user and returns a JWT token. **Requires a valid CSRF
        token** obtained from `/api/auth/csrf`. Pass the JWT token as
        `Authorization: Bearer <token>` on all protected endpoints.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, csrfToken]
              properties:
                username:
                  type: string
                  example: user1
                password:
                  type: string
                  example: password1
                csrfToken:
                  type: string
                  description: One-time token from /api/auth/csrf
                  example: a1b2c3d4e5f6...
      responses:
        "200":
          description: Login successful — use the returned token for protected endpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      username:
                        type: string
                        example: user1
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Invalid or missing CSRF token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Tasks ───────────────────────────────────────────────────────────────────
...